# perf: use an internal module resolver to improve require performance