--- a/lib/browser/api/auto-updater.js
+++ b/lib/browser/api/auto-updater.js
@@ -1,5 +1,5 @@
 if (process.platform === 'win32') {
-  module.exports = require('./auto-updater/auto-updater-win')
+  module.exports = require('@electron/internal/browser/api/auto-updater/auto-updater-win')
 } else {
-  module.exports = require('./auto-updater/auto-updater-native')
+  module.exports = require('@electron/internal/browser/api/auto-updater/auto-updater-native')
 }