--- a/lib/browser/api/auto-updater/auto-updater-win.js
+++ b/lib/browser/api/auto-updater/auto-updater-win.js
@@ -1,8 +1,6 @@
 'use strict'
 
-const { app } = require('electron')
-const { EventEmitter } = require('events')
-const squirrelUpdate = require('./squirrel-update-win')
+const squirrelUpdate = require('@electron/internal/browser/api/auto-updater/squirrel-update-win')
 
 class AutoUpdater extends EventEmitter {
   quitAndInstall () {