--- a/lib/browser/api/auto-updater/auto-updater-win.js
+++ b/lib/browser/api/auto-updater/auto-updater-win.js
@@ -1,5 +1,7 @@
 'use strict'
 
+const { app } = require('electron')
+const { EventEmitter } = require('events')
 const squirrelUpdate = require('@electron/internal/browser/api/auto-updater/squirrel-update-win')
 
 class AutoUpdater extends EventEmitter {