--- a/lib/browser/api/exports/electron.js
+++ b/lib/browser/api/exports/electron.js
@@ -8,6 +8,6 @@ common.defineProperties(exports)
 for (const module of moduleList) {
   Object.defineProperty(exports, module.name, {
     enumerable: !module.private,
-    get: () => require(`../${module.file}`)
+    get: common.memoizedGetter(() => require(`@electron/internal/browser/api/${module.file}.js`))
   })
 }