--- a/lib/browser/api/exports/electron.js
+++ b/lib/browser/api/exports/electron.js
@@ -1,6 +1,6 @@
-const common = require('../../../common/api/exports/electron')
+const common = require('@electron/internal/common/api/exports/electron')
 // since browser module list is also used in renderer, keep it separate.
-const moduleList = require('../module-list')
+const moduleList = require('@electron/internal/browser/api/module-list')
 
 // Import common modules.
 common.defineProperties(exports)