--- a/lib/browser/api/menu-item.js
+++ b/lib/browser/api/menu-item.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const roles = require('./menu-item-roles')
+const roles = require('@electron/internal/browser/api/menu-item-roles')
 
 let nextCommandId = 0