--- a/lib/browser/api/menu.js
+++ b/lib/browser/api/menu.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const { TopLevelWindow, MenuItem, webContents } = require('electron')
-const { sortMenuItems } = require('./menu-utils')
+const { sortMenuItems } = require('@electron/internal/browser/api/menu-utils')
 const EventEmitter = require('events').EventEmitter
 const v8Util = process.atomBinding('v8_util')
 const bindings = process.atomBinding('menu')