--- a/lib/browser/api/web-contents.js
+++ b/lib/browser/api/web-contents.js
@@ -6,7 +6,7 @@ const path = require('path')
 const url = require('url')
 const { app, ipcMain, session, NavigationController, deprecate } = electron
 
-const errorUtils = require('../../common/error-utils')
+const errorUtils = require('@electron/internal/common/error-utils')
 
 // session is not used here, the purpose is to make sure session is initalized
 // before the webContents module.