--- a/lib/browser/guest-view-manager.js
+++ b/lib/browser/guest-view-manager.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const { ipcMain, webContents } = require('electron')
-const parseFeaturesString = require('../common/parse-features-string')
+const parseFeaturesString = require('@electron/internal/common/parse-features-string')
 
 // Doesn't exist in early initialization.
 let webViewManager = null