--- a/lib/browser/guest-window-manager.js
+++ b/lib/browser/guest-window-manager.js
@@ -2,7 +2,7 @@
 
 const { BrowserWindow, ipcMain, webContents } = require('electron')
 const { isSameOrigin } = process.atomBinding('v8_util')
-const parseFeaturesString = require('../common/parse-features-string')
+const parseFeaturesString = require('@electron/internal/common/parse-features-string')
 
 const hasProp = {}.hasOwnProperty
 const frameToGuest = new Map()