--- a/lib/browser/init.js
+++ b/lib/browser/init.js
@@ -15,7 +15,7 @@ process.argv.splice(1, 1)
 require('../common/reset-search-paths')
 
 // Import common settings.
-require('../common/init')
+require('@electron/internal/common/init')
 
 var globalPaths = Module.globalPaths
 
@@ -94,11 +94,11 @@ if (process.platform === 'win32') {
 process.exit = app.exit
 
 // Load the RPC server.
-require('./rpc-server')
+require('@electron/internal/browser/rpc-server')
 
 // Load the guest view manager.
-require('./guest-view-manager')
-require('./guest-window-manager')
+require('@electron/internal/browser/guest-view-manager')
+require('@electron/internal/browser/guest-window-manager')
 
 // Now we try to load app's package.json.
 let packagePath = null
@@ -151,16 +151,16 @@ app.setPath('userCache', path.join(app.getPath('cache'), app.getName()))
 app.setAppPath(packagePath)
 
 // Load the chrome extension support.
-require('./chrome-extension')
+require('@electron/internal/browser/chrome-extension')
 
 const features = process.atomBinding('features')
 if (features.isDesktopCapturerEnabled()) {
   // Load internal desktop-capturer module.
-  require('./desktop-capturer')
+  require('@electron/internal/browser/desktop-capturer')
 }
 
 // Load protocol module to ensure it is populated on app ready
-require('./api/protocol')
+require('@electron/internal/browser/api/protocol')
 
 // Set main startup script of the app.
 const mainStartupScript = packageJson.main || 'index.js'