--- a/lib/browser/rpc-server.js
+++ b/lib/browser/rpc-server.js
@@ -7,8 +7,8 @@ const v8Util = process.atomBinding('v8_util')
 
 const { ipcMain, isPromise } = electron
 
-const objectsRegistry = require('./objects-registry')
-const bufferUtils = require('../common/buffer-utils')
+const objectsRegistry = require('@electron/internal/browser/objects-registry')
+const bufferUtils = require('@electron/internal/common/buffer-utils')
 
 const hasProp = {}.hasOwnProperty
 
@@ -367,13 +367,13 @@ handleRemoteCommand('ELECTRON_BROWSER_CONTEXT_RELEASE', (event, contextId) => {
 })
 
 handleRemoteCommand('ELECTRON_BROWSER_GUEST_WEB_CONTENTS', function (event, contextId, guestInstanceId) {
-  let guestViewManager = require('./guest-view-manager')
+  let guestViewManager = require('@electron/internal/browser/guest-view-manager')
   return valueToMeta(event.sender, contextId, guestViewManager.getGuest(guestInstanceId))
 })
 
 ipcMain.on('ELECTRON_BROWSER_ASYNC_CALL_TO_GUEST_VIEW', function (event, contextId, requestId, guestInstanceId, method, ...args) {
   try {
-    let guestViewManager = require('./guest-view-manager')
+    let guestViewManager = require('@electron/internal/browser/guest-view-manager')
     let guest = guestViewManager.getGuest(guestInstanceId)
     if (requestId) {
       const responseCallback = function (result) {