--- a/lib/common/api/exports/electron.js
+++ b/lib/common/api/exports/electron.js
@@ -1,4 +1,4 @@
-const moduleList = require('../module-list')
+const moduleList = require('@electron/internal/common/api/module-list')
 
 // Attaches properties to |exports|.
 exports.defineProperties = function (exports) {