--- a/lib/common/init.js
+++ b/lib/common/init.js
@@ -1,7 +1,7 @@
 const timers = require('timers')
 const util = require('util')
 
-process.atomBinding = require('./atom-binding-setup')(process.binding, process.type)
+process.atomBinding = require('@electron/internal/common/atom-binding-setup')(process.binding, process.type)
 
 // setImmediate and process.nextTick makes use of uv_check and uv_prepare to
 // run the callbacks, however since we only run uv loop on requests, the