--- a/lib/common/reset-search-paths.js
+++ b/lib/common/reset-search-paths.js
@@ -24,6 +24,9 @@ Module._nodeModulePaths = function (from) {
   }
 }
 
+const BASE_INTERNAL_PATH = path.resolve(__dirname, '..')
+const INTERNAL_MODULE_PREFIX = '@electron/internal/'
+
 // Patch Module._resolveFilename to always require the Electron API when
 // require('electron') is done.
 const electronPath = path.join(__dirname, '..', process.type, 'api', 'exports', 'electron.js')
@@ -31,6 +34,9 @@ const originalResolveFilename = Module._resolveFilename
 Module._resolveFilename = function (request, parent, isMain) {
   if (request === 'electron') {
     return electronPath
+  } else if (request.startsWith(INTERNAL_MODULE_PREFIX) && request.length > INTERNAL_MODULE_PREFIX.length) {
+    const slicedRequest = request.slice(INTERNAL_MODULE_PREFIX.length)
+    return path.resolve(BASE_INTERNAL_PATH, `${slicedRequest}${slicedRequest.endsWith('.js') ? '' : '.js'}`)
   } else {
     return originalResolveFilename(request, parent, isMain)
   }