--- a/lib/renderer/api/exports/electron.js
+++ b/lib/renderer/api/exports/electron.js
@@ -16,6 +16,6 @@ for (const {
 
   Object.defineProperty(exports, name, {
     enumerable: !isPrivate,
-    get: () => require(`../${file}`)
+    get: common.memoizedGetter(() => require(`@electron/internal/renderer/api/${file}`))
   })
 }