--- a/lib/renderer/api/exports/electron.js
+++ b/lib/renderer/api/exports/electron.js
@@ -1,5 +1,5 @@
-const common = require('../../../common/api/exports/electron')
-const moduleList = require('../module-list')
+const common = require('@electron/internal/common/api/exports/electron')
+const moduleList = require('@electron/internal/renderer/api/module-list')
 
 // Import common modules.
 common.defineProperties(exports)