--- a/lib/renderer/api/remote.js
+++ b/lib/renderer/api/remote.js
@@ -4,8 +4,8 @@ const v8Util = process.atomBinding('v8_util')
 const { ipcRenderer, isPromise } = require('electron')
 const resolvePromise = Promise.resolve.bind(Promise)
 
-const CallbacksRegistry = require('../callbacks-registry')
-const bufferUtils = require('../../common/buffer-utils')
+const CallbacksRegistry = require('@electron/internal/renderer/callbacks-registry')
+const bufferUtils = require('@electron/internal/common/buffer-utils')
 
 const callbacksRegistry = new CallbacksRegistry()
 const remoteObjectCache = v8Util.createIDWeakMap()
@@ -349,8 +349,8 @@ const addBuiltinProperty = (name) => {
 }
 
 const browserModules =
-  require('../../common/api/module-list').concat(
-    require('../../browser/api/module-list'))
+  require('@electron/internal/common/api/module-list').concat(
+    require('@electron/internal/browser/api/module-list'))
 
 // And add a helper receiver for each one.
 browserModules