--- a/lib/renderer/chrome-api.js
+++ b/lib/renderer/chrome-api.js
@@ -1,5 +1,5 @@
 const { ipcRenderer } = require('electron')
-const Event = require('./extensions/event')
+const Event = require('@electron/internal/renderer/extensions/event')
 const url = require('url')
 
 let nextId = 0
@@ -175,7 +175,7 @@ exports.injectTo = function (extensionId, isBackgroundPage, context) {
     onMessage: chrome.runtime.onMessage
   }
 
-  chrome.storage = require('./extensions/storage').setup(extensionId)
+  chrome.storage = require('@electron/internal/renderer/extensions/storage').setup(extensionId)
 
   chrome.pageAction = {
     show () {},
@@ -187,6 +187,6 @@ exports.injectTo = function (extensionId, isBackgroundPage, context) {
     getPopup () {}
   }
 
-  chrome.i18n = require('./extensions/i18n').setup(extensionId)
-  chrome.webNavigation = require('./extensions/web-navigation').setup()
+  chrome.i18n = require('@electron/internal/renderer/extensions/i18n').setup(extensionId)
+  chrome.webNavigation = require('@electron/internal/renderer/extensions/web-navigation').setup()
 }