--- a/lib/renderer/content-scripts-injector.js
+++ b/lib/renderer/content-scripts-injector.js
@@ -13,7 +13,7 @@ const matchesPattern = function (pattern) {
 // Run the code with chrome API integrated.
 const runContentScript = function (extensionId, url, code) {
   const context = {}
-  require('./chrome-api').injectTo(extensionId, false, context)
+  require('@electron/internal/renderer/chrome-api').injectTo(extensionId, false, context)
   const wrapper = `((chrome) => {\n  ${code}\n  })`
   const compiledWrapper = runInThisContext(wrapper, {
     filename: url,