--- a/lib/renderer/extensions/web-navigation.js
+++ b/lib/renderer/extensions/web-navigation.js
@@ -1,4 +1,4 @@
-const Event = require('./event')
+const Event = require('@electron/internal/renderer/extensions/event')
 const { ipcRenderer } = require('electron')
 
 class WebNavigation {