--- a/lib/renderer/web-frame-init.js
+++ b/lib/renderer/web-frame-init.js
@@ -1,5 +1,5 @@
 const { ipcRenderer, webFrame } = require('electron')
-const errorUtils = require('../common/error-utils')
+const errorUtils = require('@electron/internal/common/error-utils')
 
 module.exports = () => {
   // Call webFrame method