--- a/lib/renderer/web-view/web-view-attributes.js
+++ b/lib/renderer/web-view/web-view-attributes.js
@@ -1,7 +1,7 @@
 'use strict'
 
-const WebViewImpl = require('./web-view')
-const webViewConstants = require('./web-view-constants')
+const WebViewImpl = require('@electron/internal/renderer/web-view/web-view')
+const webViewConstants = require('@electron/internal/renderer/web-view/web-view-constants')
 const { remote } = require('electron')
 
 // Helper function to resolve url set in attribute.