--- a/lib/renderer/web-view/web-view.js
+++ b/lib/renderer/web-view/web-view.js
@@ -3,8 +3,8 @@
 const { ipcRenderer, remote, webFrame } = require('electron')
 
 const v8Util = process.atomBinding('v8_util')
-const guestViewInternal = require('./guest-view-internal')
-const webViewConstants = require('./web-view-constants')
+const guestViewInternal = require('@electron/internal/renderer/web-view/guest-view-internal')
+const webViewConstants = require('@electron/internal/renderer/web-view/web-view-constants')
 
 // An unique ID that can represent current context.
 const contextId = v8Util.getHiddenValue(global, 'contextId')