--- a/lib/sandboxed_renderer/api/exports/electron.js
+++ b/lib/sandboxed_renderer/api/exports/electron.js
@@ -1,4 +1,4 @@
-const moduleList = require('../module-list')
+const moduleList = require('@electron/internal/sandboxed_renderer/api/module-list')
 
 for (const {
   name,