--- a/lib/sandboxed_renderer/api/ipc-renderer.js
+++ b/lib/sandboxed_renderer/api/ipc-renderer.js
@@ -1,4 +1,4 @@
-const ipcRenderer = require('../../renderer/api/ipc-renderer')
+const ipcRenderer = require('@electron/internal/renderer/api/ipc-renderer')
 
 const v8Util = process.atomBinding('v8_util')
 const ipcNative = process.atomBinding('ipc')