--- a/lib/sandboxed_renderer/api/module-list.js
+++ b/lib/sandboxed_renderer/api/module-list.js
@@ -3,32 +3,32 @@ const features = process.atomBinding('features')
 module.exports = [
   {
     name: 'crashReporter',
-    load: () => require('../../common/api/crash-reporter')
+    load: () => require('@electron/internal/common/api/crash-reporter')
   },
   {
     name: 'desktopCapturer',
-    load: () => require('../../renderer/api/desktop-capturer'),
+    load: () => require('@electron/internal/renderer/api/desktop-capturer'),
     enabled: features.isDesktopCapturerEnabled()
   },
   {
     name: 'ipcRenderer',
-    load: () => require('./ipc-renderer')
+    load: () => require('@electron/internal/sandboxed_renderer/api/ipc-renderer')
   },
   {
     name: 'isPromise',
-    load: () => require('../../common/api/is-promise'),
+    load: () => require('@electron/internal/common/api/is-promise'),
     private: true
   },
   {
     name: 'nativeImage',
-    load: () => require('../../common/api/native-image')
+    load: () => require('@electron/internal/common/api/native-image')
   },
   {
     name: 'remote',
-    load: () => require('../../renderer/api/remote')
+    load: () => require('@electron/internal/renderer/api/remote')
   },
   {
     name: 'webFrame',
-    load: () => require('../../renderer/api/web-frame')
+    load: () => require('@electron/internal/renderer/api/web-frame')
   }
 ]