--- a/lib/worker/init.js
+++ b/lib/worker/init.js
@@ -11,7 +11,7 @@ process.argv.splice(1, 1)
 require('../common/reset-search-paths')
 
 // Import common settings.
-require('../common/init')
+require('@electron/internal/common/init')
 
 // Expose public APIs.
 Module.globalPaths.push(path.join(__dirname, 'api', 'exports'))