@ckerr 2018-09-15
In addition to this PR, you should also report this upstream to Babel. They'd likely appreciate the suggestion on how to improve performance to all their users.