@ckerr 2018-09-15
This is pretty cool!

Is the internal resolver needed? It looks like the memoizer would give direct improvements for every call, while the resolver would only give improvements for the first call. The benchmarks seem to back this up: the resolver gives a 30% improvement, but the memoizer is more like two orders of magnitude better.