@MarshallOfSound 2018-09-15
> Is the internal resolver needed?

The internal resolver helps with our internal require calls so all of our files, things like `error-utils` and `rpc-server` and all those files that would otherwise be required with relative paths.  It's a one-off perf improvement that isn't as good as the memoize stuff, but it's still noticeable enough to be included.  You're right though, the memoize stuff is many times more helpful 😄 Dropping init time by ~30ms is also pretty nice 😆 #smallsteps